Move Version to pack package
* Move NewPackCommand to cmd package
* Inject pack.Version into internal/builder.Save() method

* Note: The save method can really do with some refactoring, but I don't want to mix it up in this PR
